requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.28
python-json-logger>=2.0.0
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0
MetaTrader5>=5.0.45; platform_system == "Windows"
pydantic-settings>=2.0.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
